Removing original ProgramIndexView in favor of the newly redesigned UI.
