+++ server/app/views/admin/programs/ProgramIndexView.java
@@ -6,18 +6,19 @@ import static j2html.TagCreator.each;
 import static j2html.TagCreator.fieldset;
 import static j2html.TagCreator.form;
 import static j2html.TagCreator.h1;
-import static j2html.TagCreator.input;
-import static j2html.TagCreator.label;
 import static j2html.TagCreator.legend;
 import static j2html.TagCreator.p;
+import static j2html.TagCreator.span;
 
 import auth.CiviFormProfile;
+import com.google.common.collect.Lists;
 import com.google.inject.Inject;
 import com.typesafe.config.Config;
 import controllers.admin.routes;
 import j2html.tags.ContainerTag;
 import j2html.tags.Tag;
 import java.time.Instant;
+import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CompletionException;
 import play.mvc.Http;
@@ -32,21 +33,22 @@ import views.admin.AdminLayout;
 import views.admin.AdminLayout.NavPage;
 import views.admin.AdminLayoutFactory;
 import views.components.FieldWithLabel;
-import views.components.LinkElement;
+import views.components.Icons;
 import views.components.Modal;
+import views.style.AdminStyles;
 import views.style.BaseStyles;
 import views.style.ReferenceClasses;
 import views.style.StyleUtils;
 import views.style.Styles;
 
 /** Renders a page so the admin can view all active programs and draft programs. */
-public final class ProgramIndexView extends BaseHtmlView {
+public final class ProgramIndexViewV2 extends BaseHtmlView {
   private final AdminLayout layout;
   private final String baseUrl;
   private final DateConverter dateConverter;
 
   @Inject
-  public ProgramIndexView(
+  public ProgramIndexViewV2(
       AdminLayoutFactory layoutFactory, Config config, DateConverter dateConverter) {
     this.layout = checkNotNull(layoutFactory).getLayout(NavPage.PROGRAMS);
     this.baseUrl = checkNotNull(config).getString("base_url");
@@ -60,28 +62,28 @@ public final class ProgramIndexView extends BaseHtmlView {
     }
 
     String pageTitle = "All programs";
-    ContainerTag publishAllModalContent =
-        div()
-            .withClasses(Styles.FLEX, Styles.FLEX_COL, Styles.GAP_4)
-            .with(p("Are you sure you want to publish all programs?").withClasses(Styles.P_2))
-            .with(maybeRenderPublishButton(programs, request));
-    Modal publishAllModal =
-        Modal.builder("publish-all-programs-modal", publishAllModalContent)
-            .setModalTitle("Confirmation")
-            .setTriggerButtonText("Publish all programs")
-            .build();
+    Optional<Modal> maybePublishModal = maybeRenderPublishModal(programs, request);
 
     Modal demographicsCsvModal = renderDemographicsCsvModal();
     Tag contentDiv =
         div()
-            .withClasses(Styles.PX_20)
+            .withClasses(Styles.PX_4)
             .with(
-                h1(pageTitle).withClasses(Styles.MY_4),
                 div()
-                    .withClasses(Styles.FLEX, Styles.ITEMS_CENTER)
-                    .with(renderNewProgramButton())
-                    .with(div().withClass(Styles.FLEX_GROW))
-                    .condWith(programs.anyDraft(), publishAllModal.getButton()),
+                    .withClasses(
+                        Styles.FLEX,
+                        Styles.ITEMS_CENTER,
+                        Styles.SPACE_X_4,
+                        Styles.MT_12,
+                        Styles.MB_10)
+                    .with(
+                        h1(pageTitle),
+                        div().withClass(Styles.FLEX_GROW),
+                        demographicsCsvModal
+                            .getButton()
+                            .withClasses(AdminStyles.SECONDARY_BUTTON_STYLES, Styles.MY_2),
+                        renderNewProgramButton(),
+                        maybePublishModal.isPresent() ? maybePublishModal.get().getButton() : null),
                 div()
                     .withClasses(ReferenceClasses.ADMIN_PROGRAM_CARD_LIST, Styles.INVISIBLE)
                     .with(
@@ -94,23 +96,22 @@ public final class ProgramIndexView extends BaseHtmlView {
                                     programs.getActiveProgramDefinition(name),
                                     programs.getDraftProgramDefinition(name),
                                     request,
-                                    profile))))
-            .with(demographicsCsvModal.getButton());
+                                    profile))));
 
     HtmlBundle htmlBundle =
         layout
             .getBundle()
             .setTitle(pageTitle)
             .addMainContent(contentDiv)
-            .addModals(publishAllModal, demographicsCsvModal)
+            .addModals(demographicsCsvModal)
             .addFooterScripts(layout.viewUtils.makeLocalJsTag("admin_programs"));
+    maybePublishModal.ifPresent(htmlBundle::addModals);
     return layout.renderCentered(htmlBundle);
   }
 
   private Modal renderDemographicsCsvModal() {
     String modalId = "download-demographics-csv-modal";
     String downloadActionText = "Download Exported Data (CSV)";
-
     ContainerTag downloadDemographicCsvModalContent =
         div()
             .withClasses(Styles.PX_8)
@@ -140,38 +141,53 @@ public final class ProgramIndexView extends BaseHtmlView {
                                     .withClasses(Styles.ML_3, Styles.INLINE_FLEX),
                                 FieldWithLabel.date()
                                     .setFieldName("untilDate")
-                                    .setLabelText("To:")
+                                    .setLabelText("Until:")
                                     .getContainer()
                                     .withClasses(Styles.ML_3, Styles.INLINE_FLEX)),
-                        button(downloadActionText)
-                            .withClasses(BaseStyles.MODAL_BUTTON, Styles.MT_6)
+                        makeSvgTextButton(downloadActionText, Icons.DOWNLOAD)
+                            .withClasses(AdminStyles.PRIMARY_BUTTON_STYLES, Styles.MT_6)
                             .withType("submit")));
     return Modal.builder(modalId, downloadDemographicCsvModalContent)
         .setModalTitle(downloadActionText)
+        .setTriggerButtonContent(makeSvgTextButton(downloadActionText, Icons.DOWNLOAD))
         .build();
   }
 
-  private Tag maybeRenderPublishButton(ActiveAndDraftPrograms programs, Http.Request request) {
-    // We should only render the publish button if there is at least one draft.
-    if (programs.anyDraft()) {
-      String link = routes.AdminProgramController.publish().url();
-      return new LinkElement()
-          .setId("publish-programs-button")
-          .setHref(link)
-          .setText("Publish all drafts")
-          .asHiddenForm(request);
-    } else {
-      return div();
+  private ContainerTag makePublishButton() {
+    return makeSvgTextButton("Publish all drafts", Icons.PUBLISH)
+        .withId("publish-programs-button")
+        .withClasses(AdminStyles.PRIMARY_BUTTON_STYLES, Styles.MY_2);
+  }
+
+  private Optional<Modal> maybeRenderPublishModal(
+      ActiveAndDraftPrograms programs, Http.Request request) {
+    // We should only render the publish modal / button if there is at least one draft.
+    if (!programs.anyDraft()) {
+      return Optional.empty();
     }
+
+    String link = routes.AdminProgramController.publish().url();
+
+    ContainerTag publishAllModalContent =
+        div()
+            .withClasses(Styles.FLEX, Styles.FLEX_COL, Styles.GAP_4, Styles.PX_2)
+            .with(p("Are you sure you want to publish all programs?").withClasses(Styles.P_2))
+            .with(div().with(toLinkButtonForPost(makePublishButton(), link, request)));
+    Modal publishAllModal =
+        Modal.builder("publish-all-programs-modal", publishAllModalContent)
+            .setModalTitle("Confirmation")
+            .setTriggerButtonContent(makePublishButton())
+            .build();
+    return Optional.of(publishAllModal);
   }
 
   private Tag renderNewProgramButton() {
     String link = controllers.admin.routes.AdminProgramController.newOne().url();
-    return new LinkElement()
-        .setId("new-program-button")
-        .setHref(link)
-        .setText("Create new program")
-        .asButton();
+    ContainerTag button =
+        makeSvgTextButton("Create new program", Icons.ADD)
+            .withId("new-program-button")
+            .withClasses(AdminStyles.SECONDARY_BUTTON_STYLES, Styles.MY_2);
+    return asRedirectButton(button, link);
   }
 
   public ProgramDefinition getDisplayProgram(
@@ -182,82 +198,194 @@ public final class ProgramIndexView extends BaseHtmlView {
     return activeProgram.get();
   }
 
+  private Tag renderProgramRow(
+      boolean isActive,
+      ProgramDefinition program,
+      List<Tag> actions,
+      List<Tag> extraActions,
+      String... extraStyles) {
+    String badgeText = "Draft";
+    String badgeBGColor = BaseStyles.BG_CIVIFORM_PURPLE_LIGHT;
+    String badgeFillColor = BaseStyles.TEXT_CIVIFORM_PURPLE;
+    String updatedPrefix = "Edited on ";
+    Optional<Instant> updatedTime = program.lastModifiedTime();
+    if (isActive) {
+      badgeText = "Active";
+      badgeBGColor = BaseStyles.BG_CIVIFORM_GREEN_LIGHT;
+      badgeFillColor = BaseStyles.TEXT_CIVIFORM_GREEN;
+      updatedPrefix = "Published on ";
+    }
+
+    String formattedUpdateTime =
+        updatedTime.map(t -> dateConverter.renderDateTime(t)).orElse("unknown");
+    String formattedUpdateDate =
+        updatedTime.map(t -> dateConverter.renderDate(t)).orElse("unknown");
+
+    int blockCount = program.getBlockCount();
+    int questionCount = program.getQuestionCount();
+
+    String extraActionsButtonId = "extra-actions-" + program.id();
+    ContainerTag extraActionsButton =
+        makeSvgTextButton("", Icons.MORE_VERT)
+            .withId(extraActionsButtonId)
+            .withClasses(
+                AdminStyles.TERTIARY_BUTTON_STYLES,
+                ReferenceClasses.WITH_DROPDOWN,
+                Styles.H_12,
+                extraActions.size() == 0 ? Styles.INVISIBLE : "");
+
+    return div()
+        .withClasses(
+            Styles.PY_7,
+            Styles.FLEX,
+            Styles.FLEX_ROW,
+            StyleUtils.hover(Styles.BG_GRAY_100),
+            StyleUtils.joinStyles(extraStyles))
+        .with(
+            p().withClasses(
+                    badgeBGColor,
+                    badgeFillColor,
+                    Styles.ML_2,
+                    StyleUtils.responsiveXLarge(Styles.ML_8),
+                    Styles.FONT_MEDIUM,
+                    Styles.ROUNDED_FULL,
+                    Styles.FLEX,
+                    Styles.FLEX_ROW,
+                    Styles.GAP_X_2,
+                    Styles.PLACE_ITEMS_CENTER,
+                    Styles.JUSTIFY_CENTER)
+                .withStyle("min-width:90px")
+                .with(
+                    Icons.svg(Icons.NOISE_CONTROL_OFF, 20)
+                        .withClasses(Styles.INLINE_BLOCK, Styles.ML_3_5),
+                    span(badgeText).withClass(Styles.MR_4)),
+            div()
+                .withClasses(Styles.ML_4, StyleUtils.responsiveXLarge(Styles.ML_10))
+                .with(
+                    p().with(
+                            span(updatedPrefix),
+                            span(formattedUpdateTime)
+                                .withClasses(
+                                    Styles.FONT_SEMIBOLD,
+                                    Styles.HIDDEN,
+                                    StyleUtils.responsiveLarge(Styles.INLINE)),
+                            span(formattedUpdateDate)
+                                .withClasses(
+                                    Styles.FONT_SEMIBOLD,
+                                    StyleUtils.responsiveLarge(Styles.HIDDEN))),
+                    p().with(
+                            span(String.format("%d", blockCount)).withClass(Styles.FONT_SEMIBOLD),
+                            span(blockCount == 1 ? " screen, " : " screens, "),
+                            span(String.format("%d", questionCount))
+                                .withClass(Styles.FONT_SEMIBOLD),
+                            span(questionCount == 1 ? " question" : " questions"))),
+            div().withClass(Styles.FLEX_GROW),
+            div()
+                .withClasses(Styles.FLEX, Styles.SPACE_X_2, Styles.PR_6, Styles.FONT_MEDIUM)
+                .with(actions)
+                .with(
+                    div()
+                        .withClass(Styles.RELATIVE)
+                        .with(
+                            extraActionsButton,
+                            div()
+                                .withId(extraActionsButtonId + "-dropdown")
+                                .withClasses(
+                                    Styles.HIDDEN,
+                                    Styles.FLEX,
+                                    Styles.FLEX_COL,
+                                    Styles.BORDER,
+                                    Styles.BG_WHITE,
+                                    Styles.ABSOLUTE,
+                                    Styles.RIGHT_0,
+                                    Styles.W_56,
+                                    Styles.Z_50)
+                                .with(extraActions))));
+  }
+
   public Tag renderProgramListItem(
       Optional<ProgramDefinition> activeProgram,
       Optional<ProgramDefinition> draftProgram,
       Http.Request request,
       Optional<CiviFormProfile> profile) {
-    String programStatusText = extractProgramStatusText(draftProgram, activeProgram);
-
     ProgramDefinition displayProgram = getDisplayProgram(draftProgram, activeProgram);
 
-    String lastEditText =
-        displayProgram.lastModifiedTime().isPresent()
-            ? "Last updated: "
-                + dateConverter.renderDateTime(displayProgram.lastModifiedTime().get())
-            : "Could not find latest update time";
     String programTitleText = displayProgram.adminName();
     String programDescriptionText = displayProgram.adminDescription();
-    String blockCountText = "Screens: " + displayProgram.getBlockCount();
-    String questionCountText = "Questions: " + displayProgram.getQuestionCount();
 
-    Tag topContent =
-        div(
-                div(
-                    p(programStatusText).withClasses(Styles.TEXT_SM, Styles.TEXT_GRAY_700),
-                    div(programTitleText)
-                        .withClasses(
-                            ReferenceClasses.ADMIN_PROGRAM_CARD_TITLE,
-                            Styles.TEXT_BLACK,
-                            Styles.FONT_BOLD,
-                            Styles.TEXT_XL,
-                            Styles.MB_2)),
-                p().withClasses(Styles.FLEX_GROW),
-                div(p(blockCountText), p(questionCountText))
-                    .withClasses(
-                        Styles.TEXT_RIGHT,
-                        Styles.TEXT_XS,
-                        Styles.TEXT_GRAY_700,
-                        Styles.MR_2,
-                        StyleUtils.applyUtilityClass(StyleUtils.RESPONSIVE_MD, Styles.MR_4)))
-            .withClasses(Styles.FLEX);
+    ContainerTag statusDiv = div();
+    if (draftProgram.isPresent()) {
+      List<Tag> draftRowActions = Lists.newArrayList();
+      List<Tag> draftRowExtraActions = Lists.newArrayList();
+      draftRowActions.add(renderEditLink(/* isActive = */ false, draftProgram.get(), request));
+      draftRowExtraActions.add(renderManageProgramAdminsLink(draftProgram.get()));
+      draftRowExtraActions.add(renderManageTranslationsLink(draftProgram.get()));
+      statusDiv =
+          statusDiv.with(
+              renderProgramRow(
+                  /* isActive = */ false,
+                  draftProgram.get(),
+                  draftRowActions,
+                  draftRowExtraActions));
+    }
 
-    Tag midContent =
-        div(programDescriptionText)
-            .withClasses(Styles.TEXT_GRAY_700, Styles.TEXT_BASE, Styles.MB_8, Styles.LINE_CLAMP_3);
+    if (activeProgram.isPresent()) {
+      List<Tag> activeRowActions = Lists.newArrayList();
+      List<Tag> activeRowExtraActions = Lists.newArrayList();
+      Optional<Tag> applicationsLink =
+          maybeRenderViewApplicationsLink(activeProgram.get(), profile.get());
+      applicationsLink.ifPresent(activeRowExtraActions::add);
+      if (!draftProgram.isPresent()) {
+        activeRowActions.add(renderEditLink(/* isActive = */ true, activeProgram.get(), request));
+        activeRowExtraActions.add(renderManageProgramAdminsLink(activeProgram.get()));
+      }
+      activeRowActions.add(renderShareLink(activeProgram.get()));
+      statusDiv =
+          statusDiv.with(
+              renderProgramRow(
+                  /* isActive = */ true,
+                  activeProgram.get(),
+                  activeRowActions,
+                  activeRowExtraActions,
+                  draftProgram.isPresent() ? Styles.BORDER_T : ""));
+    }
 
-    Tag programDeepLink =
-        label("Deep link, use this URL to link to this program from outside of CiviForm:")
-            .withClasses(Styles.W_FULL)
+    Tag titleAndStatus =
+        div()
+            .withClass(Styles.FLEX)
             .with(
-                input()
-                    .withValue(
-                        baseUrl
-                            + controllers.applicant.routes.RedirectController.programByName(
-                                    displayProgram.slug())
-                                .url())
-                    .attr("disabled", "readonly")
-                    .withClasses(Styles.W_FULL, Styles.MB_2)
-                    .withType("text"));
+                p(programTitleText)
+                    .withClasses(
+                        ReferenceClasses.ADMIN_PROGRAM_CARD_TITLE,
+                        Styles.W_1_4,
+                        Styles.PY_7,
+                        Styles.TEXT_BLACK,
+                        Styles.FONT_BOLD,
+                        Styles.TEXT_XL),
+                statusDiv.withClasses(
+                    Styles.FLEX_GROW,
+                    Styles.TEXT_SM,
+                    StyleUtils.responsiveLarge(Styles.TEXT_BASE)));
 
-    Tag bottomContent =
-        div(
-                p(lastEditText).withClasses(Styles.TEXT_GRAY_700, Styles.ITALIC),
-                p().withClasses(Styles.FLEX_GROW),
-                maybeRenderManageTranslationsLink(draftProgram),
-                maybeRenderEditLink(draftProgram, activeProgram, request),
-                maybeRenderViewApplicationsLink(activeProgram, profile),
-                renderManageProgramAdminsLink(draftProgram, activeProgram))
-            .withClasses(Styles.FLEX, Styles.TEXT_SM, Styles.W_FULL);
-
-    Tag innerDiv =
-        div(topContent, midContent, programDeepLink, bottomContent)
-            .withClasses(
-                Styles.BORDER, Styles.BORDER_GRAY_300, Styles.BG_WHITE, Styles.ROUNDED, Styles.P_4);
-
-    return div(innerDiv)
+    return div()
         .withClasses(
-            ReferenceClasses.ADMIN_PROGRAM_CARD, Styles.W_FULL, Styles.SHADOW_LG, Styles.MB_4)
+            ReferenceClasses.ADMIN_PROGRAM_CARD,
+            Styles.W_FULL,
+            Styles.MY_4,
+            Styles.PL_6,
+            Styles.BORDER,
+            Styles.BORDER_GRAY_300,
+            Styles.ROUNDED_LG)
+        .with(
+            titleAndStatus,
+            p(programDescriptionText)
+                .withClasses(
+                    Styles.W_3_4,
+                    Styles.MB_8,
+                    Styles.PT_4,
+                    Styles.LINE_CLAMP_3,
+                    Styles.TEXT_GRAY_700,
+                    Styles.TEXT_BASE))
         // Add data attributes used for client-side sorting.
         .withData(
             "last-updated-millis",
@@ -272,118 +400,85 @@ public final class ProgramIndexView extends BaseHtmlView {
     if (draftProgram.isEmpty() && activeProgram.isEmpty()) {
       throw new IllegalArgumentException("Program neither active nor draft.");
     }
+
     ProgramDefinition program = draftProgram.isPresent() ? draftProgram.get() : activeProgram.get();
     return program.lastModifiedTime().orElse(Instant.EPOCH);
   }
 
-  private String extractProgramStatusText(
-      Optional<ProgramDefinition> draftProgram, Optional<ProgramDefinition> activeProgram) {
-    if (draftProgram.isPresent() && activeProgram.isPresent()) {
-      return "Active, with draft";
-    } else if (draftProgram.isPresent()) {
-      return "Draft";
-    } else if (activeProgram.isPresent()) {
-      return "Active";
-    }
-    throw new IllegalArgumentException("Program neither active nor draft.");
+  Tag renderShareLink(ProgramDefinition program) {
+    String programLink =
+        baseUrl
+            + controllers.applicant.routes.RedirectController.programByName(program.slug()).url();
+    return makeSvgTextButton("Share link", Icons.CONTENT_COPY)
+        .withClass(AdminStyles.TERTIARY_BUTTON_STYLES)
+        .withData("copyable-program-link", programLink);
   }
 
-  Tag maybeRenderEditLink(
-      Optional<ProgramDefinition> draftProgram,
-      Optional<ProgramDefinition> activeProgram,
-      Http.Request request) {
-    String editLinkText = "Edit →";
-    String newVersionText = "New Version";
-
-    if (draftProgram.isPresent()) {
-      String editLink =
-          controllers.admin.routes.AdminProgramController.edit(draftProgram.get().id()).url();
-
-      return new LinkElement()
-          .setId("program-edit-link-" + draftProgram.get().id())
-          .setHref(editLink)
-          .setText(editLinkText)
-          .setStyles(Styles.MR_2)
-          .asAnchorText();
-    } else if (activeProgram.isPresent()) {
-      String newVersionLink =
-          controllers.admin.routes.AdminProgramController.newVersionFrom(activeProgram.get().id())
-              .url();
-
-      return new LinkElement()
-          .setId("program-new-version-link-" + activeProgram.get().id())
-          .setHref(newVersionLink)
-          .setText(newVersionText)
-          .setStyles(Styles.MR_2)
-          .asHiddenForm(request);
-    } else {
-      // obsolete or deleted, no edit link, empty div.
-      return div();
+  Tag renderEditLink(boolean isActive, ProgramDefinition program, Http.Request request) {
+    String editLink = controllers.admin.routes.AdminProgramController.edit(program.id()).url();
+    String editLinkId = "program-edit-link-" + program.id();
+    if (isActive) {
+      editLink = controllers.admin.routes.AdminProgramController.newVersionFrom(program.id()).url();
+      editLinkId = "program-new-version-link-" + program.id();
     }
+
+    ContainerTag button =
+        makeSvgTextButton("Edit", Icons.EDIT)
+            .withId(editLinkId)
+            .withClasses(AdminStyles.TERTIARY_BUTTON_STYLES);
+    return isActive
+        ? toLinkButtonForPost(button, editLink, request)
+        : asRedirectButton(button, editLink);
   }
 
-  private Tag maybeRenderManageTranslationsLink(Optional<ProgramDefinition> draftProgram) {
-    if (draftProgram.isPresent()) {
-      String linkText = "Manage Translations →";
-      String linkDestination =
-          routes.AdminProgramTranslationsController.edit(
-                  draftProgram.get().id(), LocalizedStrings.DEFAULT_LOCALE.toLanguageTag())
-              .url();
-      return new LinkElement()
-          .setId("program-translations-link-" + draftProgram.get().id())
-          .setHref(linkDestination)
-          .setText(linkText)
-          .setStyles(Styles.MR_2)
-          .asAnchorText();
-    } else {
-      return div();
-    }
+  private Tag renderManageTranslationsLink(ProgramDefinition program) {
+    String linkDestination =
+        routes.AdminProgramTranslationsController.edit(
+                program.id(), LocalizedStrings.DEFAULT_LOCALE.toLanguageTag())
+            .url();
+    ContainerTag button =
+        makeSvgTextButton("Manage translations", Icons.LANGUAGE)
+            .withId("program-translations-link-" + program.id())
+            .withClass(AdminStyles.TERTIARY_BUTTON_STYLES);
+    return asRedirectButton(button, linkDestination);
   }
 
-  private Tag maybeRenderViewApplicationsLink(
-      Optional<ProgramDefinition> activeProgram, Optional<CiviFormProfile> userProfile) {
+  private Optional<Tag> maybeRenderViewApplicationsLink(
+      ProgramDefinition activeProgram, CiviFormProfile userProfile) {
     // TODO(#2582): Determine if this has N+1 query behavior and fix if
     // necessary.
-    if (activeProgram.isPresent() && userProfile.isPresent()) {
-      boolean userIsAuthorized = true;
-      try {
-        userProfile.get().checkProgramAuthorization(activeProgram.get().adminName()).join();
-      } catch (CompletionException e) {
-        userIsAuthorized = false;
-      }
-      if (userIsAuthorized) {
-        String editLink =
-            routes.AdminApplicationController.index(
-                    activeProgram.get().id(),
-                    Optional.empty(),
-                    Optional.empty(),
-                    Optional.empty(),
-                    Optional.empty())
-                .url();
-
-        return new LinkElement()
-            .setId("program-view-apps-link-" + activeProgram.get().id())
-            .setHref(editLink)
-            .setText("Applications →")
-            .setStyles(Styles.MR_2)
-            .asAnchorText();
-      }
+    boolean userIsAuthorized;
+    try {
+      userProfile.checkProgramAuthorization(activeProgram.adminName()).join();
+      userIsAuthorized = true;
+    } catch (CompletionException e) {
+      userIsAuthorized = false;
     }
-    return div();
+    if (userIsAuthorized) {
+      String editLink =
+          routes.AdminApplicationController.index(
+                  activeProgram.id(),
+                  Optional.empty(),
+                  Optional.empty(),
+                  Optional.empty(),
+                  Optional.empty())
+              .url();
+
+      ContainerTag button =
+          makeSvgTextButton("Applications", Icons.TEXT_SNIPPET)
+              .withId("program-view-apps-link-" + activeProgram.id())
+              .withClass(AdminStyles.TERTIARY_BUTTON_STYLES);
+      return Optional.of(asRedirectButton(button, editLink));
+    }
+    return Optional.empty();
   }
 
-  private Tag renderManageProgramAdminsLink(
-      Optional<ProgramDefinition> draftProgram, Optional<ProgramDefinition> activeProgram) {
-    // We can use the ID of either, since we just add the program name and not ID to indicate
-    // ownership.
-    long programId =
-        draftProgram.isPresent() ? draftProgram.get().id() : activeProgram.orElseThrow().id();
-    String adminLink = routes.ProgramAdminManagementController.edit(programId).url();
-    return new LinkElement()
-        .setId("manage-program-admin-link-" + programId)
-        .setHref(adminLink)
-        .setText("Manage Admins →")
-        .setStyles(Styles.MR_2)
-        .asAnchorText();
+  private Tag renderManageProgramAdminsLink(ProgramDefinition program) {
+    String adminLink = routes.ProgramAdminManagementController.edit(program.id()).url();
+    ContainerTag button =
+        makeSvgTextButton("Manage admins", Icons.GROUP)
+            .withId("manage-program-admin-link-" + program.id())
+            .withClass(AdminStyles.TERTIARY_BUTTON_STYLES);
+    return asRedirectButton(button, adminLink);
   }
 }
